### 2025-02-20 #7799 @editorialbot
[PRE REVIEW]: grepq: A Rust application that quickly filters FASTQ files by matching sequences to a set of regular expressions
**Submitting author:** @Rbfinch (<a href="http://orcid.org/0000-0002-0319-4248">Nicholas D. Crosbie</a>)
**Repository:** https://github.com/Rbfinch/grepq
**Branch with paper.md** (empty if default branch):
**Version:** v1.4.1
