### 2025-02-21 3b9e5d6
updated tests to bash bats - TAP compliant
