+++ examples/commands-4.yaml
@@ -1,23 +1,0 @@
-tests:
-  test-1: "$GREPQ 16S-iupac-and-predicates.json small.fastq"
-  test-2: "$GREPQ 16S-iupac-and-predicates.json small.fastq inverted"
-  test-3: "$GREPQ -I 16S-iupac-and-predicates.json small.fastq"
-  test-4: "$GREPQ -I 16S-iupac-and-predicates.json small.fastq inverted"
-  test-5: "$GREPQ -R 16S-iupac-and-predicates.json small.fastq"
-  test-6: "$GREPQ -R 16S-iupac-and-predicates.json small.fastq inverted"
-  test-7: "$GREPQ -c 16S-iupac-and-predicates.json small.fastq"
-  test-8: "$GREPQ -c 16S-iupac-and-predicates.json small.fastq inverted"
-  test-9: "$GREPQ 16S-iupac-and-predicates.json small.fastq tune -n 2000 -c"
-  test-10: "$GREPQ --read-gzip 16S-iupac-and-predicates.json small-copy.fastq.gz tune -n 2000 -c --names --json-matches"
-
-expected_sizes:
-  test-1: 8127
-  test-2: 445480
-  test-3: 9944
-  test-4: 545164
-  test-5: 18125
-  test-6: 993604
-  test-7: 27
-  test-8: 1480
-  test-9: 176
-  test-10: 3437
