### 2025-02-22 0efba3b
updated tests and added --summarise flag
