+++ src/lib.rs
@@ -3,6 +3,7 @@ pub mod initialise;
 pub mod inverted;
 pub mod output;
 pub mod quality;
+pub mod summarise;
 pub mod tune;
 
 #[cfg(test)]
