+++ examples/cookbook.sh
@@ -81,15 +81,27 @@ echo -e "${BLUE}\nFor each matched pattern in a search of no more than 2000 matc
 echo -e "${BLUE}\nCommand: grepq --read-gzip 16S-no-iupac.json small-copy.fastq.gz tune -n 2000 -c --names --json-matches \n${NC}"
 grepq --read-gzip 16S-no-iupac.json small-copy.fastq.gz tune -n 2000 -c --names --json-matches
 
+echo -e "${BLUE}\nAs above, but uses the summarise command to ensure that all FASTQ records are processed${NC}"
+echo -e "${BLUE}\nCommand: grepq --read-gzip 16S-no-iupac.json small-copy.fastq.gz summarise -c --names --json-matches \n${NC}"
+grepq --read-gzip 16S-no-iupac.json small-copy.fastq.gz summarise -c --names --json-matches
+
 echo -e "${BLUE}\nFor each matched pattern in a search of no more than 2000 matches of a gzip-compressed FASTQ file, print the pattern and the number of matches to a JSON file called matches.json, and include the top three most frequent
 variants of each pattern, and their respective counts${NC}"
 echo -e "${BLUE}\nCommand: grepq --read-gzip 16S-no-iupac.json small-copy.fastq.gz tune -n 2000 -c --names --json-matches --variants 3 \n${NC}"
 grepq --read-gzip 16S-no-iupac.json small-copy.fastq.gz tune -n 2000 -c --names --json-matches --variants 3
 
+echo -e "${BLUE}\nAs above, but uses the summarise command to ensure that all FASTQ records are processed${NC}"
+echo -e "${BLUE}\nCommand: grepq --read-gzip 16S-no-iupac.json small-copy.fastq.gz summarise -c --names --json-matches --variants 3 \n${NC}"
+grepq --read-gzip 16S-no-iupac.json small-copy.fastq.gz summarise -c --names --json-matches --variants 3
+
 echo -e "${BLUE}\nFor each matched pattern in a search of no more than 2000 matches of a gzip-compressed FASTQ file, print the pattern and the number of matches to a JSON file called matches.json, and include all variants of each pattern, and their respective counts. Note that the `--variants` argument is not given when `--all`
-is specified.${NC}"
+is specified${NC}"
 echo -e "${BLUE}\nCommand: grepq --read-gzip 16S-no-iupac.json small-copy.fastq.gz tune -n 2000 -c --names --json-matches --all \n${NC}"
 grepq --read-gzip 16S-no-iupac.json small-copy.fastq.gz tune -n 2000 -c --names --json-matches --all
+
+echo -e "${BLUE}\nAs above, but uses the summarise command to ensure that all FASTQ records are processed${NC}"
+echo -e "${BLUE}\nCommand: grepq --read-gzip 16S-no-iupac.json small-copy.fastq.gz summarise -c --names --json-matches --all \n${NC}"
+grepq --read-gzip 16S-no-iupac.json small-copy.fastq.gz summarise -c --names --json-matches --all
 
 echo -e "${BLUE}\nSave the records where none of the regex patterns are found${NC}"
 echo -e "${BLUE}\nCommand: grepq 16S-no-iupac.txt small.fastq inverted > out.txt \n${NC}"
