+++ book/arg.rs
@@ -4,17 +4,17 @@ use std::sync::LazyLock;
 
 static AFTER_HELP: LazyLock<String> = LazyLock::new(|| {
     format!(
-        "{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}",
+        "{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}",
         "Overview:".bold().underline(),
         "\n\n`grepq` searches the sequence line of FASTQ records for regular
 expressions that are contained in a text or JSON file, or it searches for the
 absence of those regular expressions when used with the `inverted` command. The 
 FASTQ file on which it operates can be supplied uncompressed or in gzip or zstd
-compressed format. Use the `tune` command in a simple shell script to update the
-number and order of regex patterns in your pattern file according to their matched
-frequency (refer to the examples directory of the `grepq` GitHub repository:
-https://github.com/Rbfinch/grepq), further targeting and speeding up the
-filtering process.",
+compressed format. Use the `tune` or `summarise` command in a simple shell script
+to update the number and order of regex patterns in your pattern file according
+to their matched frequency (refer to the examples directory of the `grepq` GitHub
+repository, https://github.com/Rbfinch/grepq), further targeting and speeding up
+the filtering process.",
         "\n\nExamples:".bold().underline(),
         "\n\nPrint only the matching sequences".italic(),
         "\n    grepq regex.txt file.fastq".bold(),
@@ -56,12 +56,20 @@ JSON file called matches.json"
             .italic(),
         "\n    grepq --read-gzip regex.json file.fastq.gz tune -n 100000 -c --names --json-matches"
             .bold(),
+        "\n\nAs above, but uses the summarise command to ensure that all FASTQ records are processed"
+            .italic(),
+        "\n    grepq --read-gzip regex.json file.fastq.gz summarise -c --names --json-matches"
+            .bold(),
         "\n\nFor each matched pattern in a search of no more than 100000 matches of a
 gzip-compressed FASTQ file, print the pattern and the number of matches to a 
 JSON file called matches.json, and include the top three most frequent variants of
 each pattern, and their respective counts"
             .italic(),
         "\n    grepq --read-gzip regex.json file.fastq.gz tune -n 100000 -c --names --json-matches --variants 3"
+            .bold(),
+        "\n\nAs above, but uses the summarise command to ensure that all FASTQ records are processed"
+            .italic(),
+        "\n    grepq --read-gzip regex.json file.fastq.gz summarise -c --names --json-matches --variants 3"
             .bold(),
         "\n\nFor each matched pattern in a search of no more than 100000 matches of a
 gzip-compressed FASTQ file, print the pattern and the number of matches to a JSON
@@ -70,6 +78,10 @@ respective counts. Note that the `--variants` argument is not given when `--all`
 is specified."
             .italic(),
         "\n    grepq --read-gzip regex.json file.fastq.gz tune -n 100000 -c --names --json-matches --all"
+            .bold(),
+        "\n\nAs above, but uses the summarise command to ensure that all FASTQ records are processed"
+            .italic(),
+        "\n    grepq --read-gzip regex.json file.fastq.gz summarise -c --names --json-matches --all"
             .bold(),
         "\n\nPrint the records where none of the regex patterns are found".italic(),
[…]
