+++ book/lib.rs
@@ -1,7 +1,9 @@
 pub mod arg;
 pub mod initialise;
 pub mod inverted;
+pub mod output;
 pub mod quality;
+pub mod summarise;
 pub mod tune;
 
 #[cfg(test)]
