+++ Cargo.toml
@@ -22,9 +22,9 @@ flate2 = { version = "1.0.35", features = [
     "zlib-ng",
 ], default-features = false }
 jsonschema = "0.29.0"
-serde_json = "1.0.138"
+serde_json = "1.0.139"
 colored = "3.0.0"
-zstd = "0.13.2"
+zstd = "0.13.3"
 
 [dev-dependencies]
 tempfile = "3.17.1"
