+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "grepq"
-version = "1.4.8"
+version = "1.4.9"
 edition = "2021"
 authors = ["Nicholas D. Crosbie"]
 license = "MIT"
@@ -16,11 +16,9 @@ keywords = ["bioinformatics", "FASTQ", "REGEX", "JSON", "cli"]
 seq_io = "0.3.3"
 regex = "1.11.1"
 mimalloc = "0.1.43"
-clap = { version = "4.5.30", features = ["derive", "cargo"] }
+clap = { version = "4.5.31", features = ["derive", "cargo"] }
 num_cpus = "1.16.0"
-flate2 = { version = "1.0.35", features = [
-    "zlib-ng",
-], default-features = false }
+flate2 = { version = "1.1.0", features = ["zlib-ng"], default-features = false }
 jsonschema = "0.29.0"
 serde_json = "1.0.139"
 colored = "3.0.0"
