+++ Cargo.toml
@@ -23,6 +23,8 @@ jsonschema = "0.29.0"
 serde_json = "1.0.139"
 colored = "3.0.0"
 zstd = "0.13.3"
+rusqlite = "0.33.0"
+chrono = "0.4.39"
 
 [dev-dependencies]
 tempfile = "3.17.1"
