+++ src/quality.rs
@@ -25,3 +25,32 @@ pub fn average_quality(quality: &[u8], quality_encoding: &str) -> f32 {
         0.0
     }
 }
+
+/// Calculate GC content percentage of a DNA sequence
+#[inline(always)]
+pub fn gc_content(sequence: &[u8]) -> f32 {
+    if sequence.is_empty() {
+        return 0.0;
+    }
+
+    let gc_count = sequence
+        .iter()
+        .filter(|&&base| base == b'G' || base == b'C')
+        .count();
+
+    (gc_count as f32 / sequence.len() as f32) * 100.0
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_gc_content() {
+        assert_eq!(gc_content(b"GCGC"), 100.0);
+        assert_eq!(gc_content(b"ATAT"), 0.0);
+        assert_eq!(gc_content(b"ATGC"), 50.0);
+        assert_eq!(gc_content(b""), 0.0);
+        assert_eq!(gc_content(b"GGCC"), 100.0);
+    }
+}
