+++ src/output.rs
@@ -72,7 +72,7 @@ pub fn create_sqlite_db() -> SqlResult<Connection> {
     let db_name = format!("fastq_{}.db", timestamp);
     let conn = Connection::open(&db_name)?;
 
-    // Create fastq_data table with TNF field
+    // Create fastq_data table with TNF field - use REAL instead of NUMERIC which is causing issues
     conn.execute(
         "CREATE TABLE fastq_data (
             header TEXT,
@@ -80,6 +80,8 @@ pub fn create_sqlite_db() -> SqlResult<Connection> {
             quality TEXT,
             length INTEGER,
             GC REAL,
+            GC_int INTEGER,
+            nTN INTEGER,
             TNF TEXT,
             average_quality REAL
         )",
