+++ src/lib.rs
@@ -5,6 +5,7 @@ pub mod output;
 pub mod quality;
 pub mod summarise;
 pub mod tune;
+pub mod utils;
 
 #[cfg(test)]
 mod tests;
