+++ Cargo.toml
@@ -24,7 +24,7 @@ serde_json = "1.0.139"
 colored = "3.0.0"
 zstd = "0.13.3"
 rusqlite = "0.33.0"
-chrono = "0.4.39"
+chrono = "0.4.40"
 
 [dev-dependencies]
 tempfile = "3.17.1"

[…]
