+++ Cargo.toml
@@ -25,6 +25,7 @@ colored = "3.0.0"
 zstd = "0.13.3"
 rusqlite = "0.33.0"
 chrono = "0.4.40"
+serde = "1.0.218"
 
 [dev-dependencies]
 tempfile = "3.17.1"
