+++ cookbook.md
@@ -186,3 +186,11 @@ pattern file*
 named after each regexName, with the output in FASTQ format*
 
 **grepq -R --bucket --read-gzip regex.json file.fastq.gz**
+
+<br>
+
+*For a gzip-compressed FASTQ file, bucket matched sequences into separate files
+named after each regexName, with the output in FASTQ format, and write a SQLite
+database file, limiting the number of tetranucleotides in the TNF field to two*
+
+**grepq -R --read-gzip --writeSQL -N 2 --bucket regex.json file.fastq.gz**
