### 2025-03-03 38eb630
variants JSON array
