+++ src/summarise.rs
@@ -1,5 +1,6 @@
 use crate::arg::Cli;
 use crate::initialise::{create_reader, parse_patterns_file};
+use crate::output;
 use crate::quality;
 use regex::bytes::Regex;
 use seq_io::fastq::Record;
@@ -20,6 +21,25 @@ pub fn run_summarise(cli: &Cli, include_count: bool) -> io::Result<()> {
 
     let mut match_counts: HashMap<String, usize> = HashMap::new();
     let mut match_strings: HashMap<String, HashMap<String, usize>> = HashMap::new();
+
+    // Initialize database connection if needed
+    let db_conn = if cli.write_sql {
+        let conn = if cli.patterns.ends_with(".json") {
+            let pattern_data: serde_json::Value =
+                serde_json::from_str(&std::fs::read_to_string(&cli.patterns).unwrap()).unwrap();
+            if pattern_data["regexSet"]["qualityEncoding"].is_null() {
+                output::create_sqlite_db().unwrap()
+            } else {
+                output::create_sqlite_db_with_quality().unwrap()
+            }
+        } else {
+            output::create_sqlite_db().unwrap()
+        };
+        output::write_regex_to_db(&conn, &cli.patterns, &cli.file).unwrap();
+        Some(conn)
+    } else {
+        None
+    };
 
     // Process all records
     while let Some(result) = reader.next() {
@@ -51,21 +71,61 @@ pub fn run_summarise(cli: &Cli, include_count: bool) -> io::Result<()> {
 
         // If all checks pass, match the sequence against the regex set
         if seq_len_check && header_check && qual_check {
+            let mut matches_info = vec![];
             for mat in regex_set.matches(record.seq()).into_iter() {
                 let matched_pattern = regex_set.patterns()[mat].to_string();
                 let converted_pattern = crate::initialise::convert_iupac_to_regex(&matched_pattern);
                 *match_counts.entry(converted_pattern.clone()).or_insert(0) += 1;
                 let entry = match_strings.entry(converted_pattern.clone()).or_default();
-                let matched_substring = Regex::new(&regex_set.patterns()[mat])
+                let matched = Regex::new(&regex_set.patterns()[mat])
                     .unwrap()
                     .find_iter(record.seq())
                     .next()
                     .unwrap();
-                let matched_substring =
-                    &record.seq()[matched_substring.start()..matched_substring.end()];
+                let matched_substring = &record.seq()[matched.start()..matched.end()];
                 *entry
                     .entry(String::from_utf8_lossy(matched_substring).to_string())
                     .or_insert(0) += 1;
+
+                // Collect match information
+                matches_info.push(json!({
+                    "pattern": matched_pattern,
+                    "start": matched.start(),
+                    "end": matched.end(),
+                    "match": String::from_utf8_lossy(matched_substring).to_string()
+                }));
+            }
+
+            // Calculate additional metrics
+            let avg_quality = quality_encoding
+                .as_ref()
+                .map(|encoding| quality::average_quality(record.qual(), encoding.as_str()))
+                .unwrap_or(0.0);
+            let (tnf, ntn) =
+                quality::tetranucleotide_frequencies(record.seq(), cli.num_tetranucleotides);
+            let gc = quality::gc_content(record.seq());
+            let gc_int = gc.round() as i64;
+
+            // Write matches_info to the database
+            if let Some(ref db) = db_conn {
+                let matches_json =
+                    serde_json::to_string(&matches_info).unwrap_or_else(|_| "[]".to_string());
+                db.execute(
+                    "INSERT INTO fastq_data (header, sequence, quality, length, GC, GC_int, nTN, TNF, average_quality, variants) 
[…]
