+++ src/main.rs
@@ -187,62 +187,64 @@ fn main() {
             },
             |record, found| {
                 if *found {
-                    let mut matches_info = vec![];
-                    for pattern in regex_set.patterns() {
-                        let regex = BytesRegex::new(pattern).unwrap();
-                        for matched in regex.find_iter(record.seq()) {
-                            matches_info.push(json!({
-                                "pattern": pattern,
-                                "match": String::from_utf8_lossy(&record.seq()[matched.start()..matched.end()]).to_string(),
-                                "start": matched.start(),
-                                "end": matched.end()
-                            }));
+                    if cli.write_sql {
+                        let mut matches_info = vec![];
+                        for pattern in regex_set.patterns() {
+                            let regex = BytesRegex::new(pattern).unwrap();
+                            for matched in regex.find_iter(record.seq()) {
+                                matches_info.push(json!({
+                                    "pattern": pattern,
+                                    "match": String::from_utf8_lossy(&record.seq()[matched.start()..matched.end()]).to_string(),
+                                    "start": matched.start(),
+                                    "end": matched.end()
+                                }));
+                            }
                         }
-                    }
 
-                    if let Some(ref db) = db_conn {
-                        let avg_quality = quality_encoding
-                            .map(|encoding| quality::average_quality(record.qual(), encoding))
-                            .unwrap_or(0.0);
-                        let (tnf, ntn) = quality::tetranucleotide_frequencies(record.seq(), cli.num_tetranucleotides);
-                        let gc = quality::gc_content(record.seq());
-                        let gc_int = gc.round() as i64;
-                        let matches_json = serde_json::to_string(&matches_info).unwrap_or_else(|_| "[]".to_string());
+                        if let Some(ref db) = db_conn {
+                            let avg_quality = quality_encoding
+                                .map(|encoding| quality::average_quality(record.qual(), encoding))
+                                .unwrap_or(0.0);
+                            let (tnf, ntn) = quality::tetranucleotide_frequencies(record.seq(), cli.num_tetranucleotides);
+                            let gc = quality::gc_content(record.seq());
+                            let gc_int = gc.round() as i64;
+                            let matches_json = serde_json::to_string(&matches_info).unwrap_or_else(|_| "[]".to_string());
 
-                        // Use SQLite's ROUND function to round GC and average_quality
-                        if quality_encoding.is_some() && !is_text_file {
-                            db.execute(
-                                "INSERT INTO fastq_data (header, sequence, quality, length, GC, GC_int, nTN, TNF, average_quality, variants) 
-                                 VALUES (?1, ?2, ?3, ?4, ROUND(?5, 2), ?6, ?7, ?8, ROUND(?9, 2), ?10)",
-                                rusqlite::params![
-                                    String::from_utf8_lossy(record.head()),
-                                    String::from_utf8_lossy(record.seq()),
-                                    String::from_utf8_lossy(record.qual()),
-                                    record.seq().len() as i64,
-                                    gc,
-                                    gc_int,
-                                    ntn as i64,
-                                    tnf,
-                                    avg_quality,
-                                    matches_json,
-                                ],
-                            ).unwrap();
-                        } else {
-                            db.execute(
-                                "INSERT INTO fastq_data (header, sequence, quality, length, GC, GC_int, nTN, TNF, variants) 
-                                 VALUES (?1, ?2, ?3, ?4, ROUND(?5, 2), ?6, ?7, ?8, ?9)",
-                                rusqlite::params![
-                                    String::from_utf8_lossy(record.head()),
-                                    String::from_utf8_lossy(record.seq()),
-                                    String::from_utf8_lossy(record.qual()),
-                                    record.seq().len() as i64,
-                                    gc,
-                                    gc_int,
-                                    ntn as i64,
-                                    tnf,
-                                    matches_json,
-                                ],
[…]
