+++ CHANGELOG.md
@@ -3,11 +3,18 @@ TBD
 
 Unreleased changes. Release notes have not yet been written.
 
+1.5.0
+=====
+
+* added variants field to `writeSQL` command
+* fixed summarise command performance regrssion
+* corresponding updates to documentation
+
 1.4.9
 =====
 
-**added `writeSQL` command
-** corresponding updates to documentation
+* added `writeSQL` command
+* corresponding updates to documentation
 
 1.4.8
 =====
