+++ book/initialise.rs
@@ -90,7 +90,7 @@ type ParseResult = Result<
         RegexSet,
         Option<String>,
         Option<u64>,
-        Option<f64>,
+        Option<f32>,
         Option<String>,
         Vec<String>,           // regex_names
         Vec<(String, String)>, // variants
@@ -168,7 +168,9 @@ pub fn parse_patterns_file(patterns_path: &str) -> ParseResult {
             .as_str()
             .map(|s| s.to_string());
         let minimum_sequence_length = json["regexSet"]["minimumSequenceLength"].as_u64();
-        let minimum_quality = json["regexSet"]["minimumAverageQuality"].as_f64();
+        let minimum_quality = json["regexSet"]["minimumAverageQuality"]
+            .as_f64()
+            .map(|q| q as f32);
         let quality_encoding = json["regexSet"]["qualityEncoding"]
             .as_str()
             .map(|s| s.to_string());
