+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "grepq"
-version = "1.5.1"
+version = "1.5.2"
 edition = "2021"
 authors = ["Nicholas D. Crosbie"]
 license = "MIT"
