+++ src/quality.rs
@@ -1,24 +1,27 @@
 use serde::Serialize;
 use std::collections::HashMap;
 
+// Function: average_quality
+// Calculates the average quality score for a sequence using the specified quality encoding.
 pub fn average_quality(quality: &[u8], quality_encoding: &str) -> f32 {
-    // Determine the offset based on the quality encoding
+    // Determine the offset based on the quality encoding.
+    // Phred+33: ASCII offset 33, Phred+64: ASCII offset 64.
     let offset = match quality_encoding {
         "Phred+33" => 33,
         "Phred+64" => 64,
         _ => 33, // Default to Phred+33 if unknown encoding
     };
 
-    // Initialize sum and count
+    // Initialize sum and count.
     let mut sum = 0;
     let count = quality.len() as i32;
 
-    // Calculate the sum of quality scores adjusted by the offset
+    // Iterate over quality bytes, subtracting the offset to get the true quality score.
     for &q in quality {
         sum += q as i32 - offset;
     }
 
-    // Calculate and return the average quality score
+    // Calculate the average quality score. Return 0.0 if no bases are present.
     if count > 0 {
         sum as f32 / count as f32
     } else {
@@ -26,17 +29,20 @@ pub fn average_quality(quality: &[u8], quality_encoding: &str) -> f32 {
     }
 }
 
-/// Calculate GC content percentage of a DNA sequence
-/// Only considers unambiguous bases (ACTG) in both count and total length
+/// Function: gc_content
+/// Calculates the GC content percentage of a DNA sequence.
+/// Only counts unambiguous bases (A, C, T, G); ambiguous bases are ignored.
 #[inline(always)]
 pub fn gc_content(sequence: &[u8]) -> f32 {
     if sequence.is_empty() {
         return 0.0;
     }
 
+    // Initialize counters for GC bases and total valid bases.
     let mut gc_count = 0;
     let mut valid_base_count = 0;
 
+    // Loop over each base in the sequence.
     for &base in sequence {
         match base {
             b'G' | b'C' => {
@@ -46,21 +52,26 @@ pub fn gc_content(sequence: &[u8]) -> f32 {
             b'A' | b'T' => {
                 valid_base_count += 1;
             }
-            _ => {} // Skip ambiguous bases
+            _ => {} // Skip ambiguous or non-standard bases.
         }
     }
 
+    // Avoid division by zero; return 0.0 if no valid bases found.
     if valid_base_count == 0 {
         0.0
     } else {
+        // Calculate the percentage of GC bases among valid bases.
         (gc_count as f32 / valid_base_count as f32) * 100.0
     }
 }
 
+// Function: round_to_4_sig_figs
+// Rounds a floating-point value to 4 significant figures.
 fn round_to_4_sig_figs(value: f32) -> f32 {
     if value == 0.0 {
         return 0.0;
     }
+    // Calculate scaling factor based on the logarithm of the absolute value.
     let scale = 10.0_f32.powf(4.0 - value.abs().log10().floor());
     (value * scale).round() / scale
 }
@@ -71,47 +82,45 @@ struct TetraFrequency {
     percentage: f32,
 }
 
-/// Calculate relative frequencies of tetranucleotides in a DNA sequence
-/// Returns a tuple containing:
-/// - JSON string of tetranucleotide frequencies
-/// - Count of unique tetranucleotides
-///
-/// Only considers unambiguous bases (ACTG)
-///
+/// Function: tetranucleotide_frequencies
+/// Calculates relative frequencies of tetranucleotides in a DNA sequence.
+/// Returns a tuple with a JSON string of sorted tetranucleotide frequencies and the total count of unique tetranucleotides.
+/// Only considers windows containing unambiguous bases (A, C, T, G).
[…]
