+++ src/summarise.rs
@@ -1,3 +1,7 @@
+// This module handles the summarisation command for grepq.
+// It parses the patterns file, iterates over FASTQ records, applies filters,
+// collects match statistics, writes to SQL if enabled, and prints summary output.
+
 use crate::arg::Cli;
 use crate::initialise::{create_reader, parse_patterns_file};
 use crate::output;
@@ -12,21 +16,25 @@ use std::io::{self};
 pub fn run_summarise(cli: &Cli, include_count: bool) -> io::Result<()> {
     let patterns_path = &cli.patterns;
 
-    // Parse the patterns file
-    let (regex_set, header_regex, minimum_sequence_length, minimum_quality, quality_encoding, _, _) =
-        parse_patterns_file(patterns_path).map_err(|e| io::Error::new(io::ErrorKind::Other, e))?;
+    // Parse the patterns file and extract settings.
+    let (regex_set, header_regex, minimum_sequence_length, minimum_quality, quality_encoding, _, _) = 
+        parse_patterns_file(patterns_path)
+            .map_err(|e| io::Error::new(io::ErrorKind::Other, e))?;
 
+    // Compile header regex if provided.
     let header_regex = header_regex.map(|re| Regex::new(&re).unwrap());
+    
+    // Create a reader to stream input FASTQ records.
     let mut reader = create_reader(cli);
 
+    // Initialize counters to store match counts and sub-match frequencies.
     let mut match_counts: HashMap<String, usize> = HashMap::new();
     let mut match_strings: HashMap<String, HashMap<String, usize>> = HashMap::new();
 
-    // Initialize database connection if needed
+    // Initialize database connection if SQL output is enabled.
     let db_conn = if cli.write_sql {
         let conn = if cli.patterns.ends_with(".json") {
-            let pattern_data: serde_json::Value =
-                serde_json::from_str(&std::fs::read_to_string(&cli.patterns).unwrap()).unwrap();
+            let pattern_data: serde_json::Value = serde_json::from_str(&std::fs::read_to_string(&cli.patterns).unwrap()).unwrap();
             if pattern_data["regexSet"]["qualityEncoding"].is_null() {
                 output::create_sqlite_db().unwrap()
             } else {
@@ -41,54 +49,42 @@ pub fn run_summarise(cli: &Cli, include_count: bool) -> io::Result<()> {
         None
     };
 
-    // Process all records
+    // Process each FASTQ record in a loop.
     while let Some(result) = reader.next() {
+        // Attempt to read a FASTQ record.
         let record = match result {
             Ok(record) => record,
             Err(e) => {
-                return Err(io::Error::new(
-                    io::ErrorKind::Other,
-                    format!(
-                        "grepq only supports the fastq format. Check your input file.: {}",
-                        e
-                    ),
-                ));
+                // Return error if file format is not supported.
+                return Err(io::Error::new(io::ErrorKind::Other, format!("grepq only supports the fastq format. Check your input file.: {}", e)));
             }
         };
 
-        // Check sequence length, header, and quality
-        let seq_len_check =
-            minimum_sequence_length.map_or(true, |len| record.seq().len() >= len as usize);
-        let header_check = header_regex
-            .as_ref()
-            .map_or(true, |re| re.is_match(record.head()));
+        // Apply filters: sequence length, header pattern, and quality.
+        let seq_len_check = minimum_sequence_length.map_or(true, |len| record.seq().len() >= len as usize);
+        let header_check = header_regex.as_ref().map_or(true, |re| re.is_match(record.head()));
         let qual_check = minimum_quality.map_or(true, |min_q| {
-            quality::average_quality(
-                record.qual(),
[…]
