+++ README.md
@@ -275,8 +275,6 @@ command was: ../target/release/grepq -c 16S-no-iupac.txt small.fastq <br>
 
 Further, you can run the `cookbook.sh` script in the `examples` directory to test the cookbook examples, and you can use `predate` (<https://crates.io/crates/predate>) if you prefer a Rust application to a shell script.
 
-Finally, you can run the `test.bats` script in the `test` directory to test the `grepq` executable (`/target/release/grepq`). The `bats` shell script testing framework is required to run the `test.bats` script. You can install `bats` using `brew install bats-core` on macOS, or `sudo apt-get install bats` on Ubuntu. You will need to put the the file `SRX26365298.fastq.gz` (obtain from the SRA using `fastq-dump --accession SRX26365298`) in the `examples` directory to run all tests. There is a `run_tests.sh` script in the `test` directory that will run the `test.bats` script, with various options.
-
 ```bash
 
 **SARS-CoV-2 example**
