+++ book.toml
@@ -1,6 +1,0 @@
-[book]
-authors = ["Nicholas D. Crosbie"]
-language = "en"
-multilingual = false
-src = "src"
-title = "grepq - quickly filter FASTQ files"
