+++ book/initialise.rs
@@ -10,6 +10,7 @@ use std::io::BufWriter;
 use std::io::{self, BufRead, BufReader, Write};
 use zstd::stream::{read::Decoder as ZstdDecoder, write::Encoder as ZstdEncoder};
 
+// Static JSON schema used to validate the input patterns file.
 static SCHEMA: &str = r#"
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
@@ -98,29 +99,45 @@ type ParseResult = Result<
     String,
 >;
 
-// IUAPC to regex conversion (note, DNA only)
+// Function: convert_iupac_to_regex
+// Converts IUPAC nucleotide codes to a regex pattern by replacing ambiguous characters.
+// It panics if an illegal character is encountered.
 pub fn convert_iupac_to_regex(pattern: &str) -> String {
-    let legal_chars = "ACGTYRWSKMBDHVN";
-    for c in pattern.chars() {
-        if c.is_alphabetic() && !legal_chars.contains(c.to_ascii_uppercase()) {
-            panic!("Illegal character found in pattern: {}", c);
+    let legal = "ACGTYRWSKMBDHVN";
+
+    // Check if the pattern *could* be an IUPAC pattern (all chars are alphabetic).
+    let could_be_iupac = pattern.chars().all(|c| c.is_alphabetic());
+
+    if could_be_iupac {
+        // If it *could* be IUPAC, check for illegal characters and panic if found.
+        for c in pattern.chars() {
+            if c.is_alphabetic() && !legal.contains(c.to_ascii_uppercase()) {
+                panic!("Illegal character found in pattern: {}", c);
+            }
         }
+
+        // If it's a valid IUPAC pattern, convert it to uppercase and replace characters.
+        let uppercase = pattern.to_uppercase();
+        uppercase
+            .replace('Y', "[CT]")
+            .replace('R', "[AG]")
+            .replace('W', "[AT]")
+            .replace('S', "[CG]")
+            .replace('K', "[GT]")
+            .replace('M', "[AC]")
+            .replace('B', "[CGT]")
+            .replace('D', "[AGT]")
+            .replace('H', "[ACT]")
+            .replace('V', "[ACG]")
+            .replace('N', "[ACGT]")
+    } else {
+        // If it's not IUPAC, return the original pattern unchanged.
+        pattern.to_string()
     }
-    pattern
-        .to_uppercase()
-        .replace('Y', "[CT]")
-        .replace('R', "[AG]")
-        .replace('W', "[AT]")
-        .replace('S', "[CG]")
-        .replace('K', "[GT]")
-        .replace('M', "[AC]")
-        .replace('B', "[CGT]")
-        .replace('D', "[AGT]")
-        .replace('H', "[ACT]")
-        .replace('V', "[ACG]")
-        .replace('N', "[ACGT]")
 }
 
+// Function: validate_dna_sequence
+// Validates that a given sequence contains only valid DNA nucleotides.
 fn validate_dna_sequence(sequence: &str) -> Result<(), String> {
     if sequence.chars().all(|c| "ACTG".contains(c)) {
         Ok(())
@@ -129,9 +146,12 @@ fn validate_dna_sequence(sequence: &str) -> Result<(), String> {
     }
 }
 
-// Parse patterns file (JSON or plain text)
+// Function: parse_patterns_file
+// Parses a patterns file which may be JSON or plain text, validates it against a schema,
+// and compiles regex patterns. It also extracts optional settings.
 pub fn parse_patterns_file(patterns_path: &str) -> ParseResult {
     if patterns_path.ends_with(".json") {
+        // Open and parse the JSON file.
         let json_file =
             File::open(patterns_path).map_err(|e| format!("Failed to open JSON file: {}", e))?;
[…]

[…]
