+++ paper/paper.bib
@@ -2,7 +2,8 @@
   title={Representationof Events in Nerve Nets and Finite Automata},
   author={Kleene, SC},
   journal={CE Shannon and J. McCarthy},
-  year={1951}
+  year={1951},
+  doi={10.1515/9781400882618-002}
 }
 
 @article{hodgman2000historical,
@@ -13,7 +14,8 @@
   number={1},
   pages={10--15},
   year={2000},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={doi.org/10.1093/bioinformatics/16.1.10}
 }
 
 @article{bachurin2024structural,
@@ -23,7 +25,8 @@
   volume={692},
   pages={149349},
   year={2024},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.bbrc.2023.149349}
 }
 
 @article{valdivia2012biodefense,
@@ -34,7 +37,8 @@
   number={1},
   pages={1000113},
   year={2012},
-  publisher={NIH Public Access}
+  publisher={NIH Public Access},
+  doi={10.4172/2157-2526.1000113}
 }
 
 @article{xylogiannopoulos2021pattern,
@@ -43,7 +47,8 @@
   journal={bioRxiv},
   pages={2021--04},
   year={2021},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  doi={10.1101/2021.04.14.439840}
 }
 
 @article{choi2018wastewater,
@@ -53,7 +58,8 @@
   volume={105},
   pages={453--469},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.trac.2018.06.004}
 }
 
 @article{sims2020future,
@@ -63,7 +69,8 @@
   volume={139},
   pages={105689},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.envint.2020.105689}
 }
 
 @manual{gnugrep,
@@ -140,7 +147,7 @@ urldate = {2025-01-08}
 author = {Peter, David},
 license = {MIT},
 month = mar,
-title = {{hyperfine: A command-line benchmarking tool}},
+title = {hyperfine: A command-line benchmarking tool},
 url = {https://github.com/sharkdp/hyperfine},
 version = {1.18.0},
 year = {2024}
@@ -150,34 +157,34 @@ year = {2024}
   title        = {The stat command},
   author       = {{Apple Inc.}},
   year         = {2023},
-  note         = {\url{https://ss64.com/osx/stat.html}},
+  url          = {https://ss64.com/osx/stat.html}
 }
 
 @manual{diff_macos,
   title        = {The diff command},
   author       = {{Apple Inc.}},
   year         = {2023},
-  note         = {\url{https://ss64.com/osx/diff.html}},
+  url          = {https://ss64.com/osx/diff.html}
 }
 
 @manual{top_macos,
   title        = {The top command},
   author       = {{Apple Inc.}},
   year         = {2023},
-  note         = {\url{https://ss64.com/osx/top.html}},
+  url          = {https://ss64.com/osx/top.html}
 }
 
 @manual{top_gzip,
   title        = {The gzip command},
   author       = {{Apple Inc.}},
   year         = {2019},
-  note         = {\url{https://ss64.com/osx/gzip.html}},
+  url          = {https://ss64.com/osx/gzip.html},
 }
 
 @software{spikeq,
 author = {Crosbie, Nicholas D.},
 license = {MIT},
[…]
