+++ paper/paper.md
@@ -10,7 +10,7 @@ authors:
 affiliations:
   - index: 1
     name: Melbourne Veterinary School, University of Melbourne, Parkville, Victoria, Australia
-date: "22 February 2025"
+date: "15 March 2025"
 bibliography: paper.bib
 tags: 
   - FASTQ records
@@ -22,7 +22,7 @@ tags:
 
 # Summary
 
-Regular expressions (regex) [@kleene1951representationof] have been an important tool for finding patterns in biological codes for decades [@hodgman2000historical and citations therein], and unlike fuzzy-finding approaches, do not result in approximate matches. The performance of regular expressions can be slow, however, especially when searching for matching patterns in large files. *grepq* is a Rust application that quickly filters FASTQ files by matching sequences to a set of regular expressions. *grepq* is designed with a focus on performance and scalability, is easy to install and easy to use, enabling users to quickly filter large FASTQ files, to enumerate named and unnamed variants and update the order in which patterns are matched against sequences through in-built *tune* and *summarise* commands. *grepq* is open-source and available on *GitHub* and *Crates.io*.
+Regular expressions (regex) [@kleene1951representationof] have been an important tool for finding patterns in biological codes for decades [@hodgman2000historical and citations therein], and unlike fuzzy-finding approaches, do not result in approximate matches. The performance of regular expressions can be slow, however, especially when searching for matching patterns in large files. *grepq* is a Rust application that quickly filters FASTQ files by matching sequences to a set of regular expressions. *grepq* is designed with a focus on performance and scalability, is easy to install and easy to use, enabling users to quickly filter large FASTQ files, to enumerate named and unnamed variants, to update the order in which patterns are matched against sequences through in-built *tune* and *summarise* commands, and optionally, to output a SQLite file for further sequence analysis. *grepq* is open-source and available on *GitHub*, *Crates.io* and *bioconda*.
 
 # Statement of need
 
@@ -47,6 +47,8 @@ Further performance gains were obtained by:
 - in-lining of performance-critical functions
 - use of the *write_all* I/O operation that ensures the data is written in one go, rather than writing data in smaller chunks
 
+\newpage
+
 # Feature set
 
 *grepq* has the following features:
@@ -55,13 +57,16 @@ Further performance gains were obtained by:
 - IUPAC ambiguity code support (N, R, Y, etc.)
 - support for gzip and zstd compression (reading and writing)
 - JSON support for pattern file input and *tune* and *summarise* command output, allowing named regular expression sets and named regular expressions (pattern files can also be in plain text)
-- the ability to set predicates to filter FASTQ records on the header field (= record ID line) using a regular expression, minimum sequence length, and minimum average quality score (supports Phred+33 and Phred+64)
-- the ability to output matched sequences to one of four formats (including FASTQ and FASTA)
-- the ability to tune the pattern file and enumerate named and unnamed variants with the *tune* and *summarise* commands: these commands will output a plain text or JSON file with the patterns sorted by their frequency of occurrence in the input FASTQ file or gzip-compressed FASTQ file (or for a user-specified number of total matches). This can be useful for optimizing the pattern file for performance, for example by removing patterns that are rarely matched and reordering nucleotides within the variable regions of the patterns to improve matching efficiency
-- the ability to count and summarise the total number of records and the number of matching records (or records that don't match in the case of inverted matching) in the input FASTQ file
-- the ability to bucket matching sequences to separate files named after each regexName with the `--bucket` flag, in any of the four output formats
+- the ability to:
+  - set predicates to filter FASTQ records on the header field (= record ID line) using a regular expression, minimum sequence length, and minimum average quality score (supports Phred+33 and Phred+64)
+  - output matched sequences to one of four formats (including FASTQ and FASTA)
[…]
