### 2025-03-15 #7799 @Rbfinch
> [@Rbfinch](https://github.com/Rbfinch) Dear author, thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Before we proceed, please can you have a look at the following points:
>
> * [ ]  Please study the above reference check ☝️ and see if you can address any of the reported potential DOI issues. You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again.[ ]  Your project appears to lack contributing guidelines at the moment. Please work to add these and to mention/link to them in your README. You could for instance create a `CONTRIBUTING.md` file (see here for some examples: [contributing.md/example](https://contributing.md/example/)).
>
> Lastly, we currently have a backlog of submissions to process and several editors in this domain are already at capacity. Hence I will waitlist this submission for now. We aim to assign an editor as soon as one becomes available.

Thanks.  I've regenerated the paper to account for updates and added missing DOI values where known.  Hopefully this paper can be assigned to an editor soon(ish)?
