+++ docs/tune.rs
@@ -1,204 +1,0 @@
-use crate::arg::Cli;
-use crate::initialise::{create_reader, parse_patterns_file};
-use crate::quality;
-use regex::bytes::Regex;
-use seq_io::fastq::Record;
-use serde_json::json;
-use std::collections::HashMap;
-use std::fs::File;
-use std::io::{self};
-
-// Main function to run the tune command
-pub fn run_tune(cli: &Cli, num_matches: usize, include_count: bool) -> io::Result<()> {
-    let patterns_path = &cli.patterns;
-
-    // Parse the patterns file
-    let (regex_set, header_regex, minimum_sequence_length, minimum_quality, quality_encoding, _) =
-        parse_patterns_file(patterns_path).map_err(|e| io::Error::new(io::ErrorKind::Other, e))?;
-
-    let header_regex = header_regex.map(|re| Regex::new(&re).unwrap());
-    let mut reader = create_reader(cli);
-
-    let mut match_counts: HashMap<String, usize> = HashMap::new();
-    let mut match_strings: HashMap<String, HashMap<String, usize>> = HashMap::new();
-    let mut total_matches = 0;
-
-    // Iterate through each record in the reader
-    while let Some(result) = reader.next() {
-        let record = result.map_err(|e| {
-            io::Error::new(
-                io::ErrorKind::Other,
-                format!(
-                    "grepq only supports the fastq format. Check your input file.: {}",
-                    e
-                ),
-            )
-        })?;
-
-        // Check sequence length, header, and quality
-        let seq_len_check =
-            minimum_sequence_length.map_or(true, |len| record.seq().len() >= len as usize);
-        let header_check = header_regex
-            .as_ref()
-            .map_or(true, |re| re.is_match(record.head()));
-        let qual_check = minimum_quality.map_or(true, |min_q| {
-            quality::average_quality(
-                record.qual(),
-                quality_encoding.as_deref().unwrap_or("Phred+33"),
-            ) >= min_q as f32
-        });
-
-        // If all checks pass, match the sequence against the regex set
-        if seq_len_check && header_check && qual_check {
-            for mat in regex_set.matches(record.seq()).into_iter() {
-                let matched_pattern = regex_set.patterns()[mat].to_string();
-                let converted_pattern = crate::initialise::convert_iupac_to_regex(&matched_pattern);
-                *match_counts.entry(converted_pattern.clone()).or_insert(0) += 1;
-                let entry = match_strings.entry(converted_pattern.clone()).or_default();
-                let matched_substring = Regex::new(&regex_set.patterns()[mat])
-                    .unwrap()
-                    .find_iter(record.seq())
-                    .next()
-                    .unwrap();
-                let matched_substring =
-                    &record.seq()[matched_substring.start()..matched_substring.end()];
-                *entry
-                    .entry(String::from_utf8_lossy(matched_substring).to_string())
-                    .or_insert(0) += 1;
-                total_matches += 1;
-                if total_matches >= num_matches {
-                    break;
-                }
-            }
-        }
-        if total_matches >= num_matches {
-            break;
-        }
-    }
-
-    let mut match_counts: Vec<_> = match_counts.into_iter().collect();
-    match_counts.sort_by(|a, b| b.1.cmp(&a.1));
-
-    // Handle JSON patterns file
-    if patterns_path.ends_with(".json") {
-        let json: serde_json::Value = serde_json::from_reader(std::fs::File::open(patterns_path)?)?;
-        let regex_set_name = json["regexSet"]["regexSetName"]
-            .as_str()
-            .unwrap_or("Unknown");
-
-        if cli.command.as_ref().map_or(
-            false,
[…]
