+++ docs/arg.rs
@@ -1,285 +1,0 @@
-use clap::Parser;
-use colored::*;
-use std::sync::LazyLock;
-
-static AFTER_HELP: LazyLock<String> = LazyLock::new(|| {
-    format!(
-        "{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}{}",
-        "Overview:".bold().underline(),
-        "\n\n`grepq` searches the sequence line of FASTQ records for regular
-expressions that are contained in a text or JSON file, or it searches for the
-absence of those regular expressions when used with the `inverted` command. The 
-FASTQ file on which it operates can be supplied uncompressed or in gzip or zstd
-compressed format. Use the `tune` command in a simple shell script to update the
-number and order of regex patterns in your pattern file according to their matched
-frequency (refer to the examples directory of the `grepq` GitHub repository:
-https://github.com/Rbfinch/grepq), further targeting and speeding up the
-filtering process.",
-        "\n\nExamples:".bold().underline(),
-        "\n\nPrint only the matching sequences".italic(),
-        "\n    grepq regex.txt file.fastq".bold(),
-        "\n\nPrint the matching sequences with the record ID".italic(),
-        "\n    grepq -I regex.txt file.fastq".bold(),
-        "\n\nPrint the matching sequences in FASTA format".italic(),
-        "\n    grepq -F regex.txt file.fastq".bold(),
-        "\n\nPrint the matching sequences in FASTQ format".italic(),
-        "\n    grepq -R regex.txt file.fastq".bold(),
-        "\n\nSave the matching sequences in gzip compressed FASTQ format".italic(),
-        "\n    grepq -R --write-gzip regex.txt file.fastq > output.fastq.gz".bold(),
-        "\n\nRead the FASTQ file in gzip compressed format".italic(),
-        "\n    grepq --read-gzip regex.txt file.fastq.gz".bold(),
-        "\n\nRead and save the output in gzip compressed format, with fast compression"
-            .italic(),
-        "\n    grepq --read-gzip --write-gzip --fast regex.txt file.fastq.gz > output.fastq.gz"
-            .bold(),
-        "\n\nRead and save the output in gzip compressed format, with best compression"
-            .italic(),
-        "\n    grepq --read-gzip --write-gzip --best regex.txt file.fastq.gz > output.fastq.gz"
-            .bold(),
-        "\n\nRead and save the output in zstd compressed format, with best compression"
-            .italic(),
-        "\n    grepq --read-zstd --write-zstd --best regex.txt file.fastq.zst > output.fastq.zst"
-            .bold(),
-        "\n\nCount the number of matching FASTQ records".italic(),
-        "\n    grepq -c regex.txt file.fastq".bold(),
-        "\n\nFor each matched pattern in a search of the first 100000 records, print the
-pattern and the number of matches"
-            .italic(),
-        "\n    grepq regex.txt file.fastq tune -n 100000 -c".bold(),
-        "\n\nFor each matched pattern in a search of the first 100000 records of a
-gzip-compressed FASTQ file, print the pattern and the number of matches"
-            .italic(),
-        "\n    grepq --read-gzip regex.txt file.fastq.gz tune -n 100000 -c".bold(),
-        "\n\nFor each matched pattern in a search of the first 100000 records of a
-gzip-compressed FASTQ file, print the pattern and the number of matches to a 
-JSON file called matches.json"
-            .italic(),
-        "\n    grepq --read-gzip regex.json file.fastq.gz tune -n 100000 -c --names --json-matches"
-            .bold(),
[…]
