+++ src/SUMMARY.md
@@ -1,12 +1,0 @@
-# Summary
-
-* [Feature set](Feature-set.md)
-* [Features and performance in detail](Features-and-performance-in-detail.md)
-* [Usage](Usage.md)
-* [Requirements](Requirements.md)
-* [Installation](Installation.md)
-* [Examples and tests](Examples-and-tests.md)
-* [Further testing](Further-testing.md)
-* [Citation and preprint](Citation-and-preprint.md)
-* [Update changes](Update-changes.md)
-* [License](License.md)

[…]
