+++ docs/tests.rs
@@ -1,76 +1,0 @@
-#[cfg(test)]
-mod test_module {
-    use crate::initialise;
-    use crate::quality;
-    use std::io::Write;
-    use tempfile::NamedTempFile;
-
-    #[test]
-    fn test_quality_scores() {
-        assert_eq!(quality::average_quality(b"IIIII", "Phred+33"), 40.0);
-        assert_eq!(quality::average_quality(b"hhhhh", "Phred+64"), 40.0);
-        assert_eq!(quality::average_quality(b"", "Phred+33"), 0.0);
-    }
-
-    #[test]
-    fn test_iupac_conversion() {
-        assert_eq!(initialise::convert_iupac_to_regex("ACTG"), "ACTG");
-        assert_eq!(initialise::convert_iupac_to_regex("N"), "[ACGT]");
-        assert_eq!(initialise::convert_iupac_to_regex("Y"), "[CT]");
-        assert_eq!(initialise::convert_iupac_to_regex("R"), "[AG]");
-        assert_eq!(initialise::convert_iupac_to_regex("W"), "[AT]");
-        assert_eq!(initialise::convert_iupac_to_regex("S"), "[CG]");
-        assert_eq!(initialise::convert_iupac_to_regex("K"), "[GT]");
-        assert_eq!(initialise::convert_iupac_to_regex("M"), "[AC]");
-        assert_eq!(initialise::convert_iupac_to_regex("B"), "[CGT]");
-        assert_eq!(initialise::convert_iupac_to_regex("D"), "[AGT]");
-        assert_eq!(initialise::convert_iupac_to_regex("H"), "[ACT]");
-        assert_eq!(initialise::convert_iupac_to_regex("V"), "[ACG]");
-    }
-
-    // #[test]
-    // #[should_panic(expected = "Invalid IUPAC code: X")]
-    // fn test_invalid_iupac() {
-    //     initialise::convert_iupac_to_regex("X");
-    // }
-
-    #[test]
-    fn test_pattern_parsing() {
-        let temp_file = NamedTempFile::new().unwrap();
-        let patterns = "[AG]AAT[AT]G[AG]CGGGG\n\
-                       CCCCG[CT]C[AT]ATT[CT]\n\
-                       GG[AG][ACGT]GGC[ACGT]GCAG\n\
-                       CTGC[ACGT]GCC[ACTG][CT]CC\n\
-                       G[CT][CT]G[CT]CGTCAGC\n\
-                       GCTGACG[AG]C[AG][AG]C\n\
-                       C[ACG]GC[ACGT]GC[CT]GCGG\n\
-                       CCGC[AG]GC[AGCT]GC[CGT]G\n\
-                       TAGA[AT]ACCC[ACGT][ACGT]G\n\
-                       C[ACGT][ACGT]GGGT[AT]TCTA\n\
-                       CGAGCGCAACCC\n\
-                       GGGTTGCGCTCG\n\
-                       AGG[CT]GGGGA[CT]GA\n\
-                       TC[AG]TCCCC[AG]CCT\n\
-                       [GC][CT]GGCG[ACGT]ACGGG\n\
-                       CCCGT[ACGT]CGCC[AG][GC]\n\
-                       GA[AG]GAACCTTAC\n\
-                       GTAAGGTTC[CT]TC\n\
-                       GTGGTTTAATTC\n\
-                       GAATTAAACCAC\n\
-                       G[CT]AC[AT]C[AT]CCGCC\n\
-                       GGCGG[AT]G[AT]GT[AG]C\n\
-                       GC[TG]ACACACG[CT]G\n\
-                       C[AG]CGTGTGT[AC]GC\n\
-                       G[AC]GGTGAAAT[TG]C\n\
[…]
