+++ book/lib.rs
@@ -1,10 +1,0 @@
-pub mod arg;
-pub mod initialise;
-pub mod inverted;
-pub mod output;
-pub mod quality;
-pub mod summarise;
-pub mod tune;
-
-#[cfg(test)]
-mod tests;
