### 2025-03-22 fbe39d3
Create FUNDING.yml
Signed-off-by: Nicholas D. Crosbie <milesforjazz@gmail.com>
