### 2025-04-14 #7799 @lpantano
Hello @Rbfinch , I will help with this paper. Thank you for submitting this tool to JOSS. Can you work on fixing the references? There are two with some issues, see above. My next step is to find you reviewers. If you have someone in mind let me know. Thanks!
