### 2025-04-14 #8048 @editorialbot
[REVIEW]: grepq: A Rust application that quickly filters FASTQ files by matching sequences to a set of regular expressions
**Submitting author:** @Rbfinch (<a href="http://orcid.org/0000-0002-0319-4248">Nicholas D. Crosbie</a>)
**Repository:** https://github.com/Rbfinch/grepq
**Branch with paper.md** (empty if default branch):
**Version:** grepq-v1.6.2
