+++ CHANGELOG.md
@@ -1,288 +1,241 @@
-TBD
-===
 
-Unreleased changes. Release notes have not yet been written.
+# Changelog
+
+1.5.5 (maintenance release)
+
+* updated dependencies
+* fixed typo in README.md
 
 1.5.4
-=====
 
 * improvement to command-line parsing when `writeSQL` is invoked
 * updates to documentation
 
 1.5.3
-=====
 
 * updates to documentation
 * addition of two help scripts for SQLite output, `summarise.sql` and `variants-as-json-array.sql`
 
 1.5.2
-=====
 
 * updates to documentation
 
 1.5.1
-=====
 
 * updated testing scripts
 * added export_fastq.sql to examples directory
 * corresponding updates to documentation
 
 1.5.0
-=====
 
 * added variants field to `writeSQL` command
-* fixed summarise command performance regrssion
+* fixed summarise command performance regression
 * corresponding updates to documentation
 
 1.4.9
-=====
 
 * added `writeSQL` command
 * corresponding updates to documentation
 
 1.4.8
-=====
 
 * added `summarise`command
 * corresponding updates to documentation
 * add bash bats (TAP-compliant) testing harness
 
 1.4.7
-=====
 
 * added support to bucket matched regex to files
 * fixed bug in `tune` subcommand that caused inflated count output
 * documentation updates
 
 1.4.6
-=====
 
 * documentation updates
 
 1.4.5
-=====
 
 * updates to several dependencies
 * documentation updates
 
 1.4.4
-=====
 
 * Support for named variants
 * Corresponding changes documentation
 
 1.4.3
-=====
 
-* Addtion of the --all option to output all variants to matches.json
+* Addition of the --all option to output all variants to matches.json
 * Corresponding changes documentation
 
 1.4.2
-=====
 
 * Addition of variants output to matches.json
 * Corresponding changes documentation
 
 1.4.1
-=====
 
 * Updates to README.md
 
 1.4.0
-=====
 
 * Inclusion of unit tests and benchmarks via cfg(test) & criterion
 * Updates to README.md, --help, cookbook.sh and cookbook.md
 
 1.3.9
-=====
 
 * Minor updates to README.md
 
 1.3.8
-=====
 
 * support for zstd compression
 * corresponding changes to README.md,
 --help, cookbook.sh and cookbook.md
 
 1.3.7
-=====
 
 * Minor updates to README.md
 
 1.3.6
-=====
 
 * Faster gzip support
 * Minor updates to README.md
 
 1.3.5
-=====
 
 * Support for FASTA output
 * Minor updates to test.sh and README.md
 
 1.3.4
-=====
 
 * Updates to README.md
 
 1.3.3
-=====
 
 * Minor improvements to README.md
 
 1.3.2
-=====
 
 * Minor improvements to README.md
 
 1.3.1
-=====
 
 * Minor improvements to help text and README.md
 
 1.3.0
-=====
 
 * Minor change to README.md
 
 1.2.9
-=====
 
 * Updates to README.md
 * Updates to cookbook.md
 * Addition of cookbook.sh
 
 1.2.8
-=====
 
 * Addition of cookbook.md
 * Updates to README.md
 
 1.2.7
-=====
 
 * Support for IUPAC ambiguity codes
 * Updates to README.md and help text
 * Fixed bug in `tune` command that caused incorrect output
 
 1.2.6
-=====
 
 * Updates to README.md
 
 1.2.5
-=====
 
[…]
