+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "grepq"
-version = "1.5.4"
+version = "1.5.5"
 edition = "2021"
 authors = ["Nicholas D. Crosbie"]
 license = "MIT"
@@ -15,15 +15,15 @@ keywords = ["bioinformatics", "FASTQ", "REGEX", "JSON", "cli"]
 [dependencies]
 seq_io = "0.3.4"
 regex = "1.11.1"
-mimalloc = "0.1.44"
-clap = { version = "4.5.32", features = ["derive", "cargo"] }
+mimalloc = "0.1.46"
+clap = { version = "4.5.37", features = ["derive", "cargo"] }
 num_cpus = "1.16.0"
-flate2 = { version = "1.1.0", features = ["zlib-ng"], default-features = false }
-jsonschema = "0.29.0"
+flate2 = { version = "1.1.1", features = ["zlib-ng"], default-features = false }
+jsonschema = "0.30.0"
 serde_json = "1.0.140"
 colored = "3.0.0"
 zstd = "0.13.3"
-rusqlite = "0.34.0"
+rusqlite = "0.35.0"
 chrono = "0.4.40"
 serde = "1.0.219"
 clap-markdown = "0.1.4"
