+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "grepq"
-version = "1.5.6"
+version = "1.5.7"
 edition = "2021"
 authors = ["Nicholas D. Crosbie"]
 license = "MIT"
