+++ .github/worflows/test.yml
@@ -23,6 +23,11 @@ jobs:
           toolchain: stable
           override: true
 
+      - name: Install system dependencies
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y libsqlite3-dev libzstd-dev pkg-config
+
       - name: Install hyperfine
         run: cargo install hyperfine
 
