+++ src/main.rs
@@ -25,6 +25,8 @@ use regex::bytes::Regex as BytesRegex; // Alias to avoid confusion
 // This file handles CLI argument parsing, SQL database connection (if enabled),
 // and dispatches commands to the appropriate module functions.
 
+// adding a comment to simulate a change
+
 fn main() {
     // Parse command line arguments using clap.
     let cli = Cli::parse();
