### 2025-04-26 02be2f2
Create test.yml
Signed-off-by: Nicholas D. Crosbie <milesforjazz@gmail.com>
