### 2025-04-26 c4c7f9c
Update test.yml
Signed-off-by: Nicholas D. Crosbie <milesforjazz@gmail.com>
