### 2025-04-27 cdf847d
Create ubuntu.yml
Signed-off-by: Nicholas D. Crosbie <milesforjazz@gmail.com>
