+++ .github/workflows/ubuntu.yml
@@ -1,0 +1,48 @@
+name: Linux-CI
+
+on:
+  push:
+    branches: [main, dev]
+  pull_request:
+    branches: [main]
+
+env:
+  CARGO_TERM_COLOR: always
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Set up Rust toolchain
+        uses: actions-rs/toolchain@v1
+        with:
+          toolchain: stable
+          override: true
+
+      - name: Install system dependencies
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y libsqlite3-dev libzstd-dev pkg-config
+
+      - name: Install hyperfine
+        run: cargo install hyperfine
+
+      - name: Set up Node.js for bats
+        uses: actions/setup-node@v4
+        with:
+          node-version: "20" # Or any LTS version
+
+      - name: Install bats
+        run: sudo npm install -g bats # Use sudo for global install on runner
+
+      - name: Build grepq
+        run: cargo build --release --verbose
+
+      - name: Run tests
+        run: |
+          cd test
+          ./run_tests.sh
