+++ src/main.rs
@@ -29,7 +29,7 @@ use regex::bytes::Regex as BytesRegex; // Alias to avoid confusion
 // adding a comment to simulate a change
 // adding a comment to simulate a change
 // adding a comment to simulate a change
-//
+// adding a comment to simulate a change
 
 fn main() {
     // Parse command line arguments using clap.
