### 2025-04-27 f35eda1
Update ubuntu.yml
Signed-off-by: Nicholas D. Crosbie <milesforjazz@gmail.com>
