+++ test/run_tests.sh
@@ -127,5 +127,21 @@ fi
 export APP
 export COMPUTE_TIMINGS
 
+# Base bats command
+BATS_CMD=("bats" "${SCRIPT_DIR}/test.bats")
+
+# Detect OS and add filter if Linux
+OS_TYPE=$(uname -s)
+if [ "$OS_TYPE" = "Linux" ]; then
+    # Regex to match tests 1-40 and 48+
+    LINUX_FILTER='^test-(([1-9]|[1-3][0-9]|40)|(4[8-9]|[5-9][0-9]|[1-9][0-9]{2,}))$'
+    BATS_CMD+=("--filter" "$LINUX_FILTER")
+    echo "Running on Linux, excluding tests 41-47."
+fi
+
+# Add any remaining arguments passed to the script
+BATS_CMD+=("${REMAINING_ARGS[@]}")
+
 # Run the tests
-exec bats "${SCRIPT_DIR}/test.bats" "${REMAINING_ARGS[@]}"
+echo "Executing: ${BATS_CMD[*]}"
+exec "${BATS_CMD[@]}"
