### 2025-05-17 #8048 @Ashastry2
@lpantano I am almost done with the review but I noticed issues with the test scripts. I installed grepq using bioconda and cloned the directory from github.
As instructed, I ran this command from the examples directory -
```
OS: macOS
CPU: arm64
version:
./test.sh: line 35: ../target/release/grepq: No such file or directory
./test.sh: line 50: declare: -A: invalid option
declare: usage: declare [-afFirtx] [-p] [name[=value] ...]
./test.sh: line 51: declare: -A: invalid option
declare: usage: declare [-afFirtx] [-p] [name[=value] ...]
./test.sh: line 57: tests[$key]: bad array subscript

Tests run:
2025-05-16 22:48:07

test-1
./test.sh: line 76: test-1: expression recursion level exceeded (error token is "test-1")
Summary:
Passing tests: 0
Failing tests: 0
```
Other than this, it works as expected. LMK if this warrants a separate issue in the target directory.
