### 2025-05-17 #8048 @Rbfinch
@Ashastry2 - the output suggests that the build-from-source installation option wasn't fully executed as the release binary (../target/release/grepq) isn't found (error reporing "no such file or directory"); which is required to install the /examples directory and to run the ./test.sh script.

Per the README.md ...

From source (will install the examples directory)

Clone the repository and cd into the grepq directory
Run **cargo build --release**
Relative to the cloned parent directory, the executable will be located in ./target/release
Make sure the executable is in your PATH or use the full path to the executable

Follow those instructions, cd into the examples directory, then run:

./test.sh commands-1.yaml; ./test.sh commands-2.yaml; ./test.sh commands-3.yaml; ./test.sh commands-4.yaml

I will add further text to the README.md to make this clearer.

Also, per the README.md ....

To run the test.sh and cookbook.sh scripts in the examples directory, you will need yq (v4.44.6 or later), gunzip and version 4 or later of bash.

Please let me know if these instructions don't resolve the issue.
