### 2025-05-17 #8048 @Ashastry2
Hi, thanks for the clarification. I re-ran the cargo build, and now the tests work. It would be helpful to explain it more in the README. It was unclear to me that when I use the conda install method and clone the directory, I still need to run the cargo build part for the tests to run.
