+++ src/tests.rs
@@ -1,3 +1,25 @@
+// MIT License
+
+// Copyright (c) 2024 - present Nicholas D. Crosbie
+
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
 // This module contains unit tests for various components of the grepq tool,
 // including quality scoring, IUPAC conversion, pattern parsing, GC content calculation,
 // and tetranucleotide frequency computation.
@@ -32,7 +54,7 @@ mod test_module {
         // Standard bases remain unchanged.
         assert_eq!(initialise::convert_iupac_to_regex("ACTG"), "ACTG");
         // Ambiguous codes are converted to their regex equivalents.
-        assert_eq!(initialise::convert_iupac_to_regex("N"), "[ACGT]");  // 'N' to [ACGT]
+        assert_eq!(initialise::convert_iupac_to_regex("N"), "[ACGT]"); // 'N' to [ACGT]
         assert_eq!(initialise::convert_iupac_to_regex("Y"), "[CT]");
         assert_eq!(initialise::convert_iupac_to_regex("R"), "[AG]");
         assert_eq!(initialise::convert_iupac_to_regex("W"), "[AT]");
@@ -102,10 +124,10 @@ mod test_module {
         // Additional comment: Testing GC content calculation with various types of sequences.
         // Test: Confirm GC content is correctly calculated.
         assert_eq!(quality::gc_content(b"GCGC"), 100.0); // All bases are G or C.
-        assert_eq!(quality::gc_content(b"ATAT"), 0.0);   // No GC bases.
-        assert_eq!(quality::gc_content(b"ATGC"), 50.0);   // Equal mix of GC and AT.
-        assert_eq!(quality::gc_content(b""), 0.0);         // Empty sequence returns 0.
-        assert_eq!(quality::gc_content(b"GGCC"), 100.0);    // All bases are G or C.
+        assert_eq!(quality::gc_content(b"ATAT"), 0.0); // No GC bases.
+        assert_eq!(quality::gc_content(b"ATGC"), 50.0); // Equal mix of GC and AT.
+        assert_eq!(quality::gc_content(b""), 0.0); // Empty sequence returns 0.
+        assert_eq!(quality::gc_content(b"GGCC"), 100.0); // All bases are G or C.
 
         // Test with ambiguous bases: only standard bases should be counted.
         assert_eq!(quality::gc_content(b"GCNGC"), 100.0);
@@ -129,7 +151,10 @@ mod test_module {
[…]
