+++ src/main.rs
@@ -1,3 +1,25 @@
+// MIT License
+
+// Copyright (c) 2024 - present Nicholas D. Crosbie
+
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
 use mimalloc::MiMalloc;
 #[global_allocator]
 static GLOBAL: MiMalloc = MiMalloc;
@@ -6,20 +28,20 @@ use regex::bytes::Regex;
 use seq_io::fastq::Record;
 use seq_io::parallel::parallel_fastq;
 use std::io::Write;
+use clap::Parser;
+use initialise::{create_reader, create_writer, parse_patterns_file};
+use std::io::{Error, ErrorKind};
+use serde_json::json;
+use regex::bytes::Regex as BytesRegex; // Alias to avoid confusion
+use arg::{Cli, Commands};
 
 mod arg;
-use arg::{Cli, Commands};
 mod initialise;
 mod inverted;
 mod output;
 mod quality;
 mod summarise;
 mod tune;
-use clap::Parser;
-use initialise::{create_reader, create_writer, parse_patterns_file};
-use std::io::{Error, ErrorKind};
-use serde_json::json;
-use regex::bytes::Regex as BytesRegex; // Alias to avoid confusion
 
 // Main entry point for the grepq tool.
 // This file handles CLI argument parsing, SQL database connection (if enabled),
