+++ src/summarise.rs
@@ -1,3 +1,25 @@
+// MIT License
+
+// Copyright (c) 2024 - present Nicholas D. Crosbie
+
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
 // This module handles the summarisation command for grepq.
 // It parses the patterns file, iterates over FASTQ records, applies filters,
 // collects match statistics, writes to SQL if enabled, and prints summary output.
@@ -17,13 +39,12 @@ pub fn run_summarise(cli: &Cli, include_count: bool) -> io::Result<()> {
     let patterns_path = &cli.patterns;
 
     // Parse the patterns file and extract settings.
-    let (regex_set, header_regex, minimum_sequence_length, minimum_quality, quality_encoding, _, _) = 
-        parse_patterns_file(patterns_path)
-            .map_err(|e| io::Error::new(io::ErrorKind::Other, e))?;
+    let (regex_set, header_regex, minimum_sequence_length, minimum_quality, quality_encoding, _, _) =
+        parse_patterns_file(patterns_path).map_err(|e| io::Error::new(io::ErrorKind::Other, e))?;
 
     // Compile header regex if provided.
     let header_regex = header_regex.map(|re| Regex::new(&re).unwrap());
-    
+
     // Create a reader to stream input FASTQ records.
     let mut reader = create_reader(cli);
 
@@ -34,7 +55,8 @@ pub fn run_summarise(cli: &Cli, include_count: bool) -> io::Result<()> {
     // Initialize database connection if SQL output is enabled.
     let db_conn = if cli.write_sql {
         let conn = if cli.patterns.ends_with(".json") {
-            let pattern_data: serde_json::Value = serde_json::from_str(&std::fs::read_to_string(&cli.patterns).unwrap()).unwrap();
+            let pattern_data: serde_json::Value =
+                serde_json::from_str(&std::fs::read_to_string(&cli.patterns).unwrap()).unwrap();
             if pattern_data["regexSet"]["qualityEncoding"].is_null() {
                 output::create_sqlite_db().unwrap()
             } else {
@@ -56,15 +78,27 @@ pub fn run_summarise(cli: &Cli, include_count: bool) -> io::Result<()> {
             Ok(record) => record,
             Err(e) => {
                 // Return error if file format is not supported.
-                return Err(io::Error::new(io::ErrorKind::Other, format!("grepq only supports the fastq format. Check your input file.: {}", e)));
+                return Err(io::Error::new(
+                    io::ErrorKind::Other,
+                    format!(
[…]
