+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "grepq"
-version = "1.5.7"
+version = "1.5.8"
 edition = "2021"
 authors = ["Nicholas D. Crosbie"]
 license = "MIT"
@@ -27,10 +27,15 @@ rusqlite = "0.35.0"
 chrono = "0.4.41"
 serde = "1.0.219"
 clap-markdown = "0.1.5"
+phf = "0.11.3"
+
+[build-dependencies]
+phf = "0.11.3"
+phf_codegen = "0.11.3"
 
 [dev-dependencies]
 tempfile = "3.20.0"
-criterion = { version = "0.5.1", features = ["html_reports"] }
+criterion = { version = "0.6.0", features = ["html_reports"] }
 
 [lib]
 name = "grepq"

[…]
