+++ src/tests.rs
@@ -327,4 +327,223 @@ mod test_module {
         // Clean up the temp file
         let _ = fs::remove_file(&json_path);
     }
+
+    // #[test]
+    // fn test_tetranucleotide_frequencies_variable_kmer_size() {
+    //     // Test: Verify that k-mer frequencies work with different sizes.
+    //     let sequence = b"ATCGATCGATCG";
+
+    //     // Test with trinucleotides (k=3)
+    //     let (frequencies_3, unique_count_3) =
+    //         quality::tetranucleotide_frequencies(sequence, Some(3));
+    //     let result_3: Vec<Value> = serde_json::from_str(&frequencies_3).unwrap();
+    //     assert_eq!(unique_count_3, 4);
+    //     assert_eq!(result_3.len(), 4);
+
+    //     // Test with pentanucleotides (k=5)
+    //     let (frequencies_5, unique_count_5) =
+    //         quality::tetranucleotide_frequencies(sequence, Some(5));
+    //     let result_5: Vec<Value> = serde_json::from_str(&frequencies_5).unwrap();
+    //     assert_eq!(unique_count_5, 4);
+    //     assert_eq!(result_5.len(), 4);
+
+    //     // Test with default value (should use k=4)
+    //     let (frequencies_default, unique_count_default) =
+    //         quality::tetranucleotide_frequencies(sequence, None);
+    //     let result_default: Vec<Value> = serde_json::from_str(&frequencies_default).unwrap();
+    //     assert_eq!(unique_count_default, 4);
+    //     assert_eq!(result_default.len(), 4);
+    // }
+
+    #[test]
+    fn test_tetranucleotide_frequencies_kmer_sorting() {
+        // Test: Ensure k-mers are sorted by decreasing frequency
+        let sequence = b"AAAAAAAAAATTTTGGGCCC";
+        let (frequencies, _) = quality::tetranucleotide_frequencies(sequence, Some(4));
+        let result: Vec<Value> = serde_json::from_str(&frequencies).unwrap();
+
+        // Check that results are sorted by decreasing percentage
+        let mut last_percentage = 100.0;
+        for entry in result {
+            let current_percentage = entry["percentage"].as_f64().unwrap();
+            assert!(
+                current_percentage <= last_percentage,
+                "Frequencies not sorted properly"
+            );
+            last_percentage = current_percentage;
+        }
+    }
+
+    #[test]
+    fn test_additional_quality_encodings() {
+        // Test Solexa quality encoding
+        assert_eq!(quality::average_quality(b"IIIII", "Solexa"), 40.0);
+
+        // Test Illumina 1.3+ encoding
+        assert_eq!(quality::average_quality(b"IIIII", "Illumina 1.3+"), 40.0);
+
+        // Test Illumina 1.5+ encoding
+        assert_eq!(quality::average_quality(b"IIIII", "Illumina 1.5+"), 40.0);
+
+        // Test with mixed quality scores
+        assert_eq!(
+            quality::average_quality(b"I@BCD", "Phred+33"),
+            (40.0 + 31.0 + 33.0 + 34.0 + 35.0) / 5.0
+        );
+
+        // Test with very low quality scores
+        assert_eq!(quality::average_quality(b"!!!!", "Phred+33"), 0.0);
+
+        // Test with very high quality scores
+        assert_eq!(quality::average_quality(b"~~~~", "Phred+33"), 93.0);
+    }
+
+    #[test]
+    fn test_gc_content_with_long_sequences() {
+        // Test with a longer sequence to ensure performance is reasonable
+        let long_sequence = b"GCGCGCGCGCGCGCGCGCGCATATATATATATATAT";
[…]
