+++ src/summarise.rs
@@ -325,7 +325,9 @@ pub fn run_summarise(cli: &Cli, include_count: bool) -> io::Result<()> {
     match_counts.sort_by(|a, b| b.1.cmp(&a.1));
 
     // Check if higher-order k-mers require --write-sql
-    if (cli.penta || cli.hexa || cli.hepta) && cli.write_sql.is_none() {
+    if (cli.pentanucleotides || cli.hexanucleotides || cli.heptanucleotides)
+        && cli.write_sql.is_none()
+    {
         eprintln!(
             "Error: --penta, --hexa, and --hepta options require --write-sql to be specified"
         );
@@ -351,15 +353,15 @@ pub fn run_summarise(cli: &Cli, include_count: bool) -> io::Result<()> {
             process_kmers(&sequences, KmerSize::Tetra, db_path, "tetranucleotides");
         }
 
-        if cli.penta {
+        if cli.pentanucleotides {
             process_kmers(&sequences, KmerSize::Penta, db_path, "pentanucleotides");
         }
 
-        if cli.hexa {
+        if cli.hexanucleotides {
             process_kmers(&sequences, KmerSize::Hexa, db_path, "hexanucleotides");
         }
 
-        if cli.hepta {
+        if cli.heptanucleotides {
             process_kmers(&sequences, KmerSize::Hepta, db_path, "heptanucleotides");
         }
     }
