+++ help.md
@@ -11,7 +11,7 @@ This document contains the help content for the `grepq` command-line program.
 
 ## `grepq`
 
-Copyright (c) 2024 - 2025 Nicholas D. Crosbie, licensed under the MIT License.
+Copyright (c) 2024 - present: Nicholas D. Crosbie, licensed under the MIT License.
 
 **Usage:** `grepq [OPTIONS] <PATTERNS> <FILE> [COMMAND]`
 
@@ -27,94 +27,96 @@ to their matched frequency (refer to the examples directory of the `grepq` GitHu
 repository, <https://github.com/Rbfinch/grepq>), further targeting and speeding up
 the filtering process.
 
-**Examples:**
+**Examples**:
 
-*Print only the matching sequences*  
+*Print only the matching sequences*
     **grepq regex.txt file.fastq**
 
-*Print the matching sequences with the record ID*  
+*Print the matching sequences with the record ID*
     **grepq -I regex.txt file.fastq**
 
-*Print the matching sequences in FASTA format*  
+*Print the matching sequences in FASTA format*
     **grepq -F regex.txt file.fastq**
 
-*Print the matching sequences in FASTQ format*  
+*Print the matching sequences in FASTQ format*
     **grepq -R regex.txt file.fastq**
 
-*Save the matching sequences in gzip compressed FASTQ format*  
+*Save the matching sequences in gzip compressed FASTQ format*
     **grepq -R --write-gzip regex.txt file.fastq > output.fastq.gz**
 
-*Read the FASTQ file in gzip compressed format*  
+*Read the FASTQ file in gzip compressed format*
     **grepq --read-gzip regex.txt file.fastq.gz**
 
-*Read and save the output in gzip compressed format, with fast compression*  
+*Read and save the output in gzip compressed format, with fast compression*
     **grepq --read-gzip --write-gzip --fast regex.txt file.fastq.gz > output.fastq.gz**
 
-*Read and save the output in gzip compressed format, with best compression*  
+*Read and save the output in gzip compressed format, with best compression*
     **grepq --read-gzip --write-gzip --best regex.txt file.fastq.gz > output.fastq.gz**
 
-*Read and save the output in zstd compressed format, with best compression*  
+*Read and save the output in zstd compressed format, with best compression*
     **grepq --read-zstd --write-zstd --best regex.txt file.fastq.zst > output.fastq.zst**
 
-*Count the number of matching FASTQ records*  
+*Count the number of matching FASTQ records*
     **grepq -c regex.txt file.fastq**
 
 *For each matched pattern in a search of no more than 100000 matches, print the
-    pattern and the number of matches*  
+    pattern and the number of matches*
     **grepq regex.txt file.fastq tune -n 100000 -c**
 
 *For each matched pattern in a search of no more than 100000 matches of a
-    gzip-compressed FASTQ file, print the pattern and the number of matches*  
+    gzip-compressed FASTQ file, print the pattern and the number of matches*
     **grepq --read-gzip regex.txt file.fastq.gz tune -n 100000 -c**
 
 *For each matched pattern in a search of no more than 100000 matches of a
 gzip-compressed FASTQ file, print the pattern and the number of matches to a
-JSON file called matches.json*  
+JSON file called matches.json*
     **grepq --read-gzip regex.json file.fastq.gz tune -n 100000 -c --names --json-matches**
 
-*As above, but uses the summarise command to ensure that all FASTQ records are processed*  
+*As above, but uses the summarise command to ensure that all FASTQ records are processed*
     **grepq --read-gzip regex.json file.fastq.gz summarise -c --names --json-matches**
 
 *For each matched pattern in a search of no more than 100000 matches of a
[…]
