+++ README.md
@@ -10,12 +10,12 @@ _Quickly filter FASTQ files_
 - [Feature set](#feature-set)
 - [Features and performance in detail](#features-and-performance-in-detail)
 - [Usage](#usage)
-- [Cookbook](https://github.com/Rbfinch/grepq/blob/main/cookbook.md)
+  - [Preparing pattern files](#preparing-pattern-files)
 - [Requirements](#requirements)
 - [Installation](#installation)
 - [Examples and tests](#examples-and-tests)
-- [Further testing](#futher-testing)
-- [Citation and preprint](#citation)
+- [Further testing](#further-testing)
+- [Citation and preprint](#citation-and-preprint)
 - [Update changes](#update-changes)
 - [Contributing](#contributing)
 - [License](#license)
@@ -32,7 +32,7 @@ This README contains documentation for the latest version of `grepq`. If you are
 - use **predicates** to filter on the header field (= record ID line) using a regex, minimum sequence length, and minimum average quality score (supports Phred+33 and Phred+64)
 - does not match false positives
 - output matched sequences to one of four formats
-- optionally output matched sequences to a **SQLite** database file, including GC content, tetranucleotide frequencies, and regex pattern matches and their position(s) in each matched FASTQ sequence, allowing for further analysis
+- optionally output matched sequences to a **SQLite** database file, including GC content, tetranucleotide and canonical tetranucleotide frequencies, and regex pattern matches and their position(s) in each matched FASTQ sequence, allowing for further analysis
 - tune your pattern file and **enumerate named and unnamed variants** with the `tune` command (use the `summarise` command to process all FASTQ records)
 - **bucket matching sequences** to separate files named after each regexName with the `--bucket` flag, in any of the four output formats
 - supports inverted matching with the `inverted` command
@@ -85,7 +85,7 @@ Predicates are specified in a JSON pattern file. For an example, see `16S-iupac-
 
 **4. Does not match false positives**
 
-`grepq` will only match regex patterns to the sequence field of a FASTQ record, which is the most common use case. Unlike `ripgrep` and `grep`, which will match the regex patterns to the entire FASTQ record, which includes the record ID, sequence, separator, and quality fields. This can lead to false positives and slow down the filtering process.
+`grepq` will only match regex patterns to the sequence of a FASTQ record, which is the most common use case. Unlike `ripgrep` and `grep`, which will match the regex patterns to the entire FASTQ record, which includes the record ID, sequence, separator, and quality fields. This can lead to false positives and slow down the filtering process.
 
 **5. Output matched sequences to one of four formats**
 
@@ -95,11 +95,11 @@ Predicates are specified in a JSON pattern file. For an example, see `16S-iupac-
 - FASTQ format (`-R` option)
 
 >[!NOTE]
-Other than when the `tune` or `summarise` command is run (see below), a FASTQ record is deemed to match (and hence provided in the output) when _any_ of the regex patterns in the pattern file match the sequence field of the FASTQ record.
+Other than when the `tune` or `summarise` command is run (see below), a FASTQ record is deemed to match (and hence provided in the output) when _any_ of the regex patterns in the pattern file match the sequence of the FASTQ record.
 
 **6. Optionally output matched sequences to a SQLite database file**
 
[…]

[…]
