+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "grepq"
-version = "1.5.8"
+version = "1.5.9"
 edition = "2021"
 authors = ["Nicholas D. Crosbie"]
 license = "MIT"
