### 2025-05-27 #8048 @lpantano
Thank you all for your work! @Rbfinch , can you reduce the paper to 3 pages? maybe some of the information can be links pointing to the repository or documentation? Let me know.
