+++ paper/paper.md
@@ -32,26 +32,9 @@ Regular expressions are a powerful tool for matching sequences, but they can be 
 
 # Implementation
 
-*grepq* is implemented in Rust, a systems programming language known for its safety features, which help prevent common programming errors such as null pointer dereferences and buffer overflows. These features make Rust an ideal choice for implementing a tool like *grepq*, which needs to be fast, efficient, and reliable.
-
-Furthermore, *grepq* obtains its performance and reliability, in part, by using the *seq_io* [@seq_io] and *regex* [@regex] libraries. The *seq_io* library is a well-tested library for parsing FASTQ files, designed to be fast and efficient, and which includes a module for parallel processing of FASTQ records through multi-threading. The *regex* library is designed to work with regular expressions and sets of regular expressions, and is known to be one of the fastest regular expression libraries currently available [@rebar]. The *regex* library supports Perl-like regular expressions without look-around or backreferences (documented at <https://docs.rs/regex/1.*/regex/#syntax>).
-
-Further performance gains were obtained by:
-
-- use of the *RegexSet* struct from the *regex* library to match multiple regular expressions against a sequence in a single pass, rather than matching each regular expression individually (the *RegexSet* is created and compiled once before entering any loop that processes the FASTQ records, avoiding the overhead of recompiling the regular expressions for each record)
-- multi-threading to process the records within an input FASTQ file in parallel through use of multiple CPU cores
-- use of the *zlib-ng* backend to the *flate2* library to read and write gzip-compressed FASTQ files, which is faster than the default *miniz_oxide* backend
-- use of an optimised global memory allocator (the *mimalloc* library [@mimalloc]) to reduce memory fragmentation and improve memory allocation and deallocation performance
-- buffer reuse to reduce the number of memory allocations and deallocations
-- use of byte slices to avoid the overhead of converting to and from string types
-- in-lining of performance-critical functions
-- use of the *write_all* I/O operation that ensures the data is written in one go, rather than writing data in smaller chunks
-
-\newpage
+*grepq* obtains its performance and reliability, in part, by using the *seq_io* [@seq_io] and *regex* [@regex] libraries. The *seq_io* library is a well-tested library for parsing FASTQ files, designed to be fast and efficient, and which includes a module for parallel processing of FASTQ records through multi-threading. The *regex* library is designed to work with regular expressions and sets of regular expressions, and is known to be one of the fastest regular expression libraries currently available [@rebar]. The *regex* library supports Perl-like regular expressions without look-around or backreferences (documented at <https://docs.rs/regex/1.*/regex/#syntax>).
 
 # Feature set
-
-*grepq* has the following features:
 
 - support for presence and absence (inverted) matching of a set of regular expressions
 - IUPAC ambiguity code support (N, R, Y, etc.)
@@ -64,122 +47,15 @@ Further performance gains were obtained by:
   - count and summarise the total number of records and the number of matching records (or records that don't match in the case of inverted matching) in the input FASTQ file
   - bucket matching sequences to separate files named after each regexName with the **--bucket** flag, in any of the four output formats
 
[…]
