+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "grepq"
-version = "1.5.9"
+version = "1.6.0"
 edition = "2021"
 authors = ["Nicholas D. Crosbie"]
 license = "MIT"
@@ -16,14 +16,14 @@ keywords = ["bioinformatics", "FASTQ", "REGEX", "JSON", "cli"]
 seq_io = "0.3.4"
 regex = "1.11.1"
 mimalloc = "0.1.46"
-clap = { version = "4.5.38", features = ["derive", "cargo"] }
-num_cpus = "1.16.0"
+clap = { version = "4.5.39", features = ["derive", "cargo"] }
+num_cpus = "1.17.0"
 flate2 = { version = "1.1.1", features = ["zlib-ng"], default-features = false }
 jsonschema = "0.30.0"
 serde_json = "1.0.140"
 colored = "3.0.0"
 zstd = "0.13.3"
-rusqlite = "0.35.0"
+rusqlite = "0.36.0"
 chrono = "0.4.41"
 serde = "1.0.219"
 clap-markdown = "0.1.5"
