+++ src/arg.rs
@@ -171,18 +171,18 @@ is supported with the `writeSQL` option. The SQLite database will contain a tabl
 called `fastq_data` with the following fields: the fastq record (header, sequence
 and quality fields), length of the sequence (length), percent GC content (GC),
 percent GC content as an integer (GC_int), number of unique tetranucleotides in the
-sequence (nTN), number of unique canonical tetranucleotides in the sequence, percent
-tetranucleotide frequency in the sequence (TNF), percent canonical tetranucleotide
-frequency in the sequence (CTNF), and a JSON array containing the matched regex
-patterns, the matches and their position(s) in the FASTQ sequence (variants). If the
-pattern file was given in JSON format and contained a non-null qualityEncoding field,
-then the average quality score for the sequence (average_quality) will also be
-written. The `--num-tetranucleotides` option can be used to limit the number of 
-tetranucleotides written to the TNF and CTNF fields of the fastq_data SQLite table,
-these being the most or equal most frequent tetranucleotides and canonical 
-tetranucleotides in the sequence of the matched FASTQ records. A summary of
-the invoked query (pattern and data files) is written to a second table called
-`query`.
+sequence (nTN), number of unique canonical tetranucleotides in the sequence (nCTN),
+percent tetranucleotide frequency in the sequence (TNF), percent canonical 
+tetranucleotide frequency in the sequence (CTNF), and a JSON array containing the
+matched regex patterns, the matches and their position(s) in the FASTQ sequence 
+(variants). If the pattern file was given in JSON format and contained a non-null
+qualityEncoding field, then the average quality score for the sequence 
+(average_quality) will also be written. The `--num-tetranucleotides` option can be
+used to limit the number of tetranucleotides written to the TNF and CTNF fields of
+the fastq_data SQLite table, these being the most or equal most frequent 
+tetranucleotides and canonical tetranucleotides in the sequence of the matched 
+FASTQ records. A summary of the invoked query (pattern and data files) is written
+to a second table called `query`.
 
 3. Pattern files must contain one regex pattern per line or be given in JSON
 format, and patterns are case-sensitive (you can supply an empty pattern file to
