### 2025-05-31 #8048 @Rbfinch
@lpantano After some editing and shifting code listings and tables to a supplemental file, the paper is now a little over 3.5 pages inclusive all references and 3 pages without references.   It'll loose coherence if I trim it more.  Is this sufficient? (see recompiled paper above).
