+++ paper/paper.bib
@@ -1,279 +1,279 @@
 @article{kleene1951representationof,
-  title={Representationof Events in Nerve Nets and Finite Automata},
-  author={Kleene, SC},
-  journal={CE Shannon and J. McCarthy},
-  year={1951},
-  doi={10.1515/9781400882618-002}
+  title   = {Representationof Events in Nerve Nets and Finite Automata},
+  author  = {Kleene, SC},
+  journal = {CE Shannon and J. McCarthy},
+  year    = {1951},
+  doi     = {10.1515/9781400882618-002}
 }
 
 @article{hodgman2000historical,
-  title={A historical perspective on gene/protein functional assignment},
-  author={Hodgman, T. Charles},
-  journal={Bioinformatics},
-  volume={16},
-  number={1},
-  pages={10--15},
-  year={2000},
-  publisher={Oxford University Press},
-  doi={doi.org/10.1093/bioinformatics/16.1.10}
+  title     = {A historical perspective on gene/protein functional assignment},
+  author    = {Hodgman, T. Charles},
+  journal   = {Bioinformatics},
+  volume    = {16},
+  number    = {1},
+  pages     = {10--15},
+  year      = {2000},
+  publisher = {Oxford University Press},
+  doi       = {10.1093/bioinformatics/16.1.10}
 }
 
 @article{bachurin2024structural,
-  title={Structural peculiarities of tandem repeats and their clinical significance},
-  author={Bachurin, Stanislav S and Yurushkin, Mikhail V and Slynko, Ilya A and Kletskii, Mikhail E and Burov, Oleg N and Berezovskiy, Dmitriy P},
-  journal={Biochemical and Biophysical Research Communications},
-  volume={692},
-  pages={149349},
-  year={2024},
-  publisher={Elsevier},
-  doi={10.1016/j.bbrc.2023.149349}
+  title     = {Structural peculiarities of tandem repeats and their clinical significance},
+  author    = {Bachurin, Stanislav S and Yurushkin, Mikhail V and Slynko, Ilya A and Kletskii, Mikhail E and Burov, Oleg N and Berezovskiy, Dmitriy P},
+  journal   = {Biochemical and Biophysical Research Communications},
+  volume    = {692},
+  pages     = {149349},
+  year      = {2024},
+  publisher = {Elsevier},
+  doi       = {10.1016/j.bbrc.2023.149349}
 }
 
 @article{valdivia2012biodefense,
-  title={Biodefense oriented genomic-based pathogen classification systems: challenges and opportunities},
-  author={Valdivia-Granda, Willy A},
-  journal={Journal of bioterrorism \& biodefense},
-  volume={3},
-  number={1},
-  pages={1000113},
-  year={2012},
-  publisher={NIH Public Access},
-  doi={10.4172/2157-2526.1000113}
+  title     = {Biodefense oriented genomic-based pathogen classification systems: challenges and opportunities},
+  author    = {Valdivia-Granda, Willy A},
+  journal   = {Journal of bioterrorism \& biodefense},
+  volume    = {3},
+  number    = {1},
+  pages     = {1000113},
+  year      = {2012},
+  publisher = {NIH Public Access},
+  doi       = {10.4172/2157-2526.1000113}
 }
 
 @article{xylogiannopoulos2021pattern,
-  title={Pattern Detection in Multiple Genome Sequences with Applications: The Case of All SARS-CoV-2 Complete Variants},
-  author={Xylogiannopoulos, Konstantinos F},
-  journal={bioRxiv},
-  pages={2021--04},
-  year={2021},
-  publisher={Cold Spring Harbor Laboratory},
[…]
