+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "grepq"
-version = "1.6.1"
+version = "1.6.2"
 edition = "2021"
 authors = ["Nicholas D. Crosbie"]
 license = "MIT"
