### 2025-06-07 #8048 @Rbfinch
@lpantano

I've completed the author tasks - see my responses to those tasks below:

Double check authors and affiliations (including ORCIDs) - Checked
Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.  - Release version is grepq-v1.6.2
Archive the release on Zenodo/figshare/etc and post the DOI here. -  Done and DOI is 10.5281/zenodo.15612061
Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper. - Checked and they match
Make sure that the license listed for the archive is the same as the software license. - Checked and they are the same

**References**

I've updated one of the references to remove a DOI prefix issue.  Note that there are no DOI for the references listed under "SKIP DOI" by the editorial bot.  Also note that the suggested DOI for "The stat command" is not appropriate as that refers to an old windows or DOS application and not a posix application that was used (which doesn't have a DOI).
