### 2025-06-10 8973d40
small fixes paper.md
Signed-off-by: Lorena Pantano <lorena.pantano@gmail.com>
