### 2025-06-20 #8048 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
