### 2025-06-21 #8048 @Rbfinch
@lpantano - I've checked the proof and it looks fine.  Note that the "missing DOI" flagged during the dry run is false.  It appears to refer to an old windows or dos command of the same name (i.e., not the commonly known posix command cited in the paper, for which there is no DOI).
