### 2025-06-27 #8048 @Kevin-Mattheus-Moerman
- [ ] `occassionally` should read `occasionally`
- [ ] `optimized` and `optimizing` should read `optimised` and `optimising` (in line with the British English used throughout)
