### 2025-06-27 b8aa03c
corrected spelling issues in paper.md
