### 2025-06-30 a160d49
Update CITATION.cff
Updated to JOSS paper citation

Signed-off-by: Nicholas D. Crosbie <milesforjazz@gmail.com>
