### 2025-06-30 19bb4c0
Update README.md
Updated README.md to point to JOSS paper and citation

Signed-off-by: Nicholas D. Crosbie <milesforjazz@gmail.com>
